[Websockets] Refactor to integrate connector
