+++ octobot_trading/exchanges/connectors/__init__.pxd
@@ -24,12 +24,11 @@ from octobot_trading.exchanges.connectors.ccxt_exchange cimport (
 )
 from octobot_trading.exchanges.connectors cimport websocket_connector
 from octobot_trading.exchanges.connectors.websocket_connector cimport (
-    WebSocketConnector,
+    WebsocketConnector,
 )
-
 
 __all__ = [
     "ExchangeSimulator",
-    "WebSocketConnector",
     "CCXTExchange",
+    "WebsocketConnector",
 ]
