+++ octobot_trading/exchanges/types/__init__.pxd
@@ -14,16 +14,8 @@
 #  You should have received a copy of the GNU Lesser General Public
 #  License along with this library.
 
-from octobot_trading.exchanges.types cimport future_exchange
-from octobot_trading.exchanges.types cimport websocket_exchange
-from octobot_trading.exchanges.types cimport margin_exchange
-from octobot_trading.exchanges.types cimport spot_exchange
-
 from octobot_trading.exchanges.types.future_exchange cimport (
     FutureExchange,
-)
-from octobot_trading.exchanges.types.websocket_exchange cimport (
-    WebsocketExchange,
 )
 from octobot_trading.exchanges.types.margin_exchange cimport (
     MarginExchange,
@@ -31,10 +23,13 @@ from octobot_trading.exchanges.types.margin_exchange cimport (
 from octobot_trading.exchanges.types.spot_exchange cimport (
     SpotExchange,
 )
+from octobot_trading.exchanges.types.websocket_exchange cimport (
+    WebSocketExchange,
+)
 
 __all__ = [
     "FutureExchange",
-    "WebsocketExchange",
+    "WebSocketExchange",
     "MarginExchange",
     "SpotExchange",
 ]
