+++ octobot_trading/exchanges/types/__init__.py
@@ -18,10 +18,6 @@ from octobot_trading.exchanges.types import spot_exchange
 from octobot_trading.exchanges.types.spot_exchange import (
     SpotExchange,
 )
-from octobot_trading.exchanges.types import websocket_exchange
-from octobot_trading.exchanges.types.websocket_exchange import (
-    WebsocketExchange,
-)
 from octobot_trading.exchanges.types import margin_exchange
 from octobot_trading.exchanges.types.margin_exchange import (
     MarginExchange,
@@ -30,10 +26,14 @@ from octobot_trading.exchanges.types import future_exchange
 from octobot_trading.exchanges.types.future_exchange import (
     FutureExchange,
 )
+from octobot_trading.exchanges.types import websocket_exchange
+from octobot_trading.exchanges.types.websocket_exchange import (
+    WebSocketExchange,
+)
 
 __all__ = [
     "FutureExchange",
-    "WebsocketExchange",
+    "WebSocketExchange",
     "MarginExchange",
     "SpotExchange",
 ]
